Fix handling debian version in prepare-chroot-* scripts
The script get requested debian version as its second parameter. Save it
to a variable actually used in the script. Previously the script used
DEBIANVERSION variable (and in some places DIST), but the parameter was
set to DISTRO variable not used anywhere else. This caused the parameter
to be ignored and previous value of the DEBIANVERSION/DIST variables
used instead (which by a coincidence were the same).
Fix this by using DIST variable in all of those places consistently.